skyfocus: refactor focus command into BlazeModule (SkyfocusModule) and implement `build --experimental_working_set=<files>`.
The new `build` flags are:

* `--experimental_enable_skyfocus`: the main feature flag.
* `--experimental_working_set`: replacement of `focus --experimental_working_set`
* `--experimental_skyfocus_dump_keys`: replacement of `focus --dump_keys`
* `--experimental_skyfocus_dump_post_gc_stats`: replacement of `focus --dump_used_heap_size_after_gc`

In addition, with the move to the `BlazeModule`, I also replaced the trigger point for Skyfocus with an EventBus `@Subscribe`r. The nice thing about that is we can easily hook Skyfocus into various points of the Blaze lifecycle, such as post-analysis for reducing peak heap in the future, instead of just post-execution for reducing retained heap.

I do acknowledge that EventBus makes debugging and tracing the control flow a bit more difficult, but that's a tradeoff I believe it's worth making for the added lifecycle control.

There is no change to the `focus` algorithm in this CL.

[1] https://github.com/google/guava/wiki/EventBusExplained#avoid-eventbus

PiperOrigin-RevId: 604620930
Change-Id: Ie0713f7ae0dc02babfcfdf7a89df5d85a1c35d89